investigator, position, protocols refactored
