+++ login.go
@@ -2,7 +2,10 @@ package main
 
 import (
 	"context"
+	"errors"
 	"fmt"
+
+	"github.com/jsandberg07/clitest/internal/database"
 )
 
 func (cfg *Config) login() error {
@@ -27,4 +30,18 @@ func (cfg *Config) login() error {
 
 func (cfg *Config) printLogin() {
 	fmt.Printf("Logged in as %s -- %s\n", cfg.loggedInInvestigator.IName, cfg.loggedInPosition.Title)
+}
+
+func (cfg *Config) getInvestigator(name string) (database.Investigator, error) {
+	investigators, err := cfg.db.GetInvestigatorByName(context.Background(), name)
+	if err != nil {
+		fmt.Println("Error getting investigator")
+		return database.Investigator{}, nil
+	}
+	if len(investigators) > 1 {
+		fmt.Println("Error getting investigator")
+		return database.Investigator{}, errors.New("vague investigator name")
+	}
+
+	return investigators[0], nil
 }
